Adds users formatting and separates TTT tests
